Move globals handling out of util/coordinator.py
